requirements.txt
anthropic>=0.84.0
boto3>=1.34.0
python-dotenv>=1.0.0
